+++ ai/core/src/main/kotlin/org/http4k/connect/model/deprecatedAiCore.kt
@@ -1,5 +1,7 @@
 package org.http4k.connect.model
 
+import org.http4k.ai.model.Prompt
+
 @Deprecated("use org.http4k.ai.model.MaxTokens")
 typealias MaxTokens = org.http4k.ai.model.MaxTokens
 
@@ -12,8 +14,8 @@ typealias Role = org.http4k.ai.model.Role
 @Deprecated("use org.http4k.ai.model.StopReason")
 typealias StopReason = org.http4k.ai.model.StopReason
 
-@Deprecated("use org.http4k.ai.model.SystemMessage")
-typealias SystemMessage = org.http4k.ai.model.SystemMessage
+@Deprecated("use org.http4k.ai.model.Prompt")
+typealias SystemMessage = Prompt
 
 @Deprecated("use org.http4k.ai.model.Temperature")
 typealias Temperature = org.http4k.ai.model.Temperature