+++ connect/ai/ollama/client/src/main/kotlin/org/http4k/connect/ollama/deprecated.kt
@@ -1,10 +1,10 @@
 package org.http4k.connect.ollama
 
-import org.http4k.ai.model.SystemMessage
+import org.http4k.ai.model.Prompt
 
-@Deprecated("use common prompt")
+@Deprecated("use org.http4k.ai.model.Prompt")
 typealias Prompt = org.http4k.ai.model.Prompt
 
-@Deprecated("use common SystemMessage")
-typealias SystemMessage = SystemMessage
+@Deprecated("use org.http4k.ai.model.Prompt")
+typealias SystemMessage = Prompt