+++ ai/core/src/main/kotlin/org/http4k/ai/model/Prompt.kt
@@ -1,8 +1,0 @@
-package org.http4k.ai.model
-
-import dev.forkhandles.values.NonBlankStringValueFactory
-import dev.forkhandles.values.StringValue
-
-class Prompt private constructor(value: String) : StringValue(value) {
-    companion object : NonBlankStringValueFactory<Prompt>(::Prompt)
-}