+++ ai/core/src/main/kotlin/org/http4k/ai/util/withAiMappings.kt
@@ -3,13 +3,14 @@ package org.http4k.ai.util
 import com.squareup.moshi.JsonAdapter
 import org.http4k.ai.model.MaxTokens
 import org.http4k.ai.model.ModelName
-import org.http4k.ai.model.Prompt
 import org.http4k.ai.model.RequestId
 import org.http4k.ai.model.ResponseId
 import org.http4k.ai.model.Role
 import org.http4k.ai.model.StopReason
+import org.http4k.ai.model.SystemPrompt
 import org.http4k.ai.model.Temperature
 import org.http4k.ai.model.ToolName
+import org.http4k.ai.model.UserPrompt
 import org.http4k.connect.model.Base64Blob
 import org.http4k.connect.model.MimeType
 import org.http4k.connect.model.Timestamp
@@ -21,7 +22,8 @@ fun <T> AutoMappingConfiguration<T>.withAiMappings() = apply {
     value(Base64Blob)
     value(MimeType)
     value(Timestamp)
-    value(Prompt)
+    value(SystemPrompt)
+    value(UserPrompt)
     value(ModelName)
     value(MaxTokens)
     value(ModelName)