+++ ai/core/src/main/kotlin/org/http4k/connect/model/deprecatedAiCore.kt
@@ -1,6 +1,6 @@
 package org.http4k.connect.model
 
-import org.http4k.ai.model.Prompt
+import org.http4k.ai.model.SystemPrompt
 
 @Deprecated("use org.http4k.ai.model.MaxTokens")
 typealias MaxTokens = org.http4k.ai.model.MaxTokens
@@ -14,8 +14,8 @@ typealias Role = org.http4k.ai.model.Role
 @Deprecated("use org.http4k.ai.model.StopReason")
 typealias StopReason = org.http4k.ai.model.StopReason
 
-@Deprecated("use org.http4k.ai.model.Prompt")
-typealias SystemMessage = Prompt
+@Deprecated("use org.http4k.ai.model.SystemPrompt")
+typealias SystemMessage = SystemPrompt
 
 @Deprecated("use org.http4k.ai.model.Temperature")
 typealias Temperature = org.http4k.ai.model.Temperature