+++ connect/ai/anthropic/client/src/main/kotlin/org/http4k/connect/anthropic/action/AbstractMessageCompletion.kt
@@ -2,7 +2,8 @@ package org.http4k.connect.anthropic.action
 
 import org.http4k.ai.model.MaxTokens
 import org.http4k.ai.model.ModelName
-import org.http4k.ai.model.Prompt
+import org.http4k.ai.model.SystemPrompt
+import org.http4k.ai.model.UserPrompt
 import org.http4k.ai.model.Temperature
 import org.http4k.connect.anthropic.ToolChoice
 
@@ -12,7 +13,7 @@ sealed interface AbstractMessageCompletion {
     val max_tokens: MaxTokens
     val metadata: Metadata?
     val stop_sequences: List<String>
-    val system: Prompt?
+    val system: SystemPrompt?
     val temperature: Temperature?
     val tool_choice: ToolChoice?
     val tools: List<Tool>