+++ connect/ai/anthropic/client/src/main/kotlin/org/http4k/connect/anthropic/action/MessageCompletion.kt
@@ -4,10 +4,11 @@ import dev.forkhandles.result4k.Failure
 import dev.forkhandles.result4k.Success
 import org.http4k.ai.model.MaxTokens
 import org.http4k.ai.model.ModelName
-import org.http4k.ai.model.Prompt
+import org.http4k.ai.model.UserPrompt
 import org.http4k.ai.model.ResponseId
 import org.http4k.ai.model.Role
 import org.http4k.ai.model.StopReason
+import org.http4k.ai.model.SystemPrompt
 import org.http4k.ai.model.Temperature
 import org.http4k.connect.Http4kConnectAction
 import org.http4k.connect.anthropic.AnthropicAIAction
@@ -29,7 +30,7 @@ data class MessageCompletion internal constructor(
     override val max_tokens: MaxTokens,
     override val metadata: Metadata? = null,
     override val stop_sequences: List<String> = emptyList(),
-    override val system: Prompt? = null,
+    override val system: SystemPrompt? = null,
     override val temperature: Temperature? = Temperature.ZERO,
     override val tool_choice: ToolChoice? = null,
     override val tools: List<Tool> = emptyList(),
@@ -39,11 +40,11 @@ data class MessageCompletion internal constructor(
 ) : AbstractMessageCompletion, AnthropicAIAction<MessageCompletionResponse> {
     constructor(
         model: ModelName,
-        prompt: Prompt,
+        prompt: UserPrompt,
         max_tokens: MaxTokens,
         metadata: Metadata? = null,
         stop_sequences: List<String> = emptyList(),
-        system: Prompt? = null,
+        system: SystemPrompt? = null,
         temperature: Temperature? = Temperature.ZERO,
         tool_choice: ToolChoice? = null,
         tools: List<Tool> = emptyList(),
@@ -61,7 +62,7 @@ data class MessageCompletion internal constructor(
         max_tokens: MaxTokens,
         metadata: Metadata? = null,
         stop_sequences: List<String> = emptyList(),
-        system: Prompt? = null,
+        system: SystemPrompt? = null,
         temperature: Temperature? = Temperature.ZERO,
         tool_choice: ToolChoice? = null,
         tools: List<Tool> = emptyList(),