+++ connect/ai/azure/client/src/main/kotlin/org/http4k/connect/azure/action/Completion.kt
@@ -1,7 +1,7 @@
 package org.http4k.connect.azure.action
 
 import org.http4k.ai.model.MaxTokens
-import org.http4k.ai.model.Prompt
+import org.http4k.ai.model.UserPrompt
 import org.http4k.ai.model.Temperature
 import org.http4k.connect.Http4kConnectAction
 import org.http4k.connect.azure.AzureAIMoshi.autoBody
@@ -13,7 +13,7 @@ import se.ansman.kotshi.JsonSerializable
 @Http4kConnectAction
 @JsonSerializable
 data class Completion(
-    val prompt: Prompt,
+    val prompt: UserPrompt,
     val presence_penalty: Double = 0.0,
     val frequency_penalty: Double = 0.0,
     val temperature: Temperature = Temperature.ONE,
@@ -27,7 +27,7 @@ data class Completion(
     override fun toRequest() = Request(POST, "/completions")
         .with(autoBody<Completion>().toLens() of this)
 
-    constructor(prompt: Prompt, max_tokens: MaxTokens, stream: Boolean = true) : this(
+    constructor(prompt: UserPrompt, max_tokens: MaxTokens, stream: Boolean = true) : this(
         prompt,
         max_tokens = max_tokens,
         stream = stream,