+++ connect/ai/azure/client/src/main/kotlin/org/http4k/connect/azure/deprecated.kt
@@ -3,7 +3,7 @@ package org.http4k.connect.azure
 import org.http4k.ai.model.ResponseId
 
 @Deprecated("use common prompt")
-typealias Prompt = org.http4k.ai.model.Prompt
+typealias Prompt = org.http4k.ai.model.UserPrompt
 
 @Deprecated("use common CompletionId")
 typealias CompletionId = ResponseId