+++ connect/ai/ollama/client/src/main/kotlin/org/http4k/connect/ollama/action/Completion.kt
@@ -1,7 +1,8 @@
 package org.http4k.connect.ollama.action
 
 import org.http4k.ai.model.ModelName
-import org.http4k.ai.model.Prompt
+import org.http4k.ai.model.SystemPrompt
+import org.http4k.ai.model.UserPrompt
 import org.http4k.connect.Http4kConnectAction
 import org.http4k.connect.model.Base64Blob
 import org.http4k.connect.ollama.OllamaAction
@@ -21,10 +22,10 @@ import java.time.Instant
 @JsonSerializable
 data class Completion(
     val model: ModelName,
-    val prompt: Prompt,
+    val prompt: UserPrompt,
     val images: List<Base64Blob>? = null,
     val stream: Boolean? = false,
-    val system: Prompt? = null,
+    val system: SystemPrompt? = null,
     val format: ResponseFormat? = null,
     val template: Template? = null,
     val raw: Boolean? = null,