+++ connect/ai/ollama/client/src/main/kotlin/org/http4k/connect/ollama/deprecated.kt
@@ -1,10 +1,11 @@
 package org.http4k.connect.ollama
 
-import org.http4k.ai.model.Prompt
+import org.http4k.ai.model.SystemPrompt
+import org.http4k.ai.model.UserPrompt
 
 @Deprecated("use org.http4k.ai.model.Prompt")
-typealias Prompt = org.http4k.ai.model.Prompt
+typealias Prompt = UserPrompt
 
-@Deprecated("use org.http4k.ai.model.Prompt")
-typealias SystemMessage = Prompt
+@Deprecated("use org.http4k.ai.model.SystemPrompt")
+typealias SystemMessage = SystemPrompt